external/Certiphy/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.3
